Matter: rename & clean up files
