+++ lib/matter_ext.scons
@@ -422,12 +422,12 @@ sources = [
     f"{MATTER_SUPPORT_DIR}/tinycrypt/src/ecc.c",
     f"{MATTER_SUPPORT_DIR}/tinycrypt/src/tinycrypt_util.c",
 
+    f"{MATTER_GLUE_DIR}/platform/bsb/BSBCommissionableDataProvider.cpp",
     f"{MATTER_GLUE_DIR}/platform/bsb/BSBDeviceInfoProvider.cpp",
+    f"{MATTER_GLUE_DIR}/platform/bsb/BSBDeviceInstanceInfoProvider.cpp",
     f"{MATTER_GLUE_DIR}/platform/bsb/CHIPCryptoPALTinyCrypt.cpp",
-    f"{MATTER_GLUE_DIR}/platform/bsb/CommissionableDataProviderImpl.cpp",
     f"{MATTER_GLUE_DIR}/platform/bsb/ConfigurationManagerImpl.cpp",
     f"{MATTER_GLUE_DIR}/platform/bsb/ConnectivityManagerImpl.cpp",
-    f"{MATTER_GLUE_DIR}/platform/bsb/DeviceInstanceInfoProviderImpl.cpp",
     f"{MATTER_GLUE_DIR}/platform/bsb/DiagnosticDataProviderImpl.cpp",
     f"{MATTER_GLUE_DIR}/platform/bsb/GenericPlatformManagerImplFuri.cpp",
     f"{MATTER_GLUE_DIR}/platform/bsb/KeyValueStoreManagerImpl.cpp",
