+++ lib/matter_glue/platform/bsb/BSBDeviceInstanceInfoProvider.cpp
@@ -1,4 +1,4 @@
-#include "DeviceInstanceInfoProviderImpl.hpp"
+#include "BSBDeviceInstanceInfoProvider.hpp"
 
 #include <platform/CHIPDeviceConfig.h>
