+++ lib/matter_glue/platform/bsb/PlatformManagerImpl.h
@@ -26,15 +26,6 @@
 
 #include "GenericPlatformManagerImplFuri.h"
 
-#if CHIP_DEVICE_CONFIG_ENABLE_WIFI_STATION
-// #include <platform/silabs/wifi/WifiInterfaceAbstraction.h>
-#endif
-#include <cmsis_os2.h>
-
-#if CHIP_DEVICE_CONFIG_ENABLE_WIFI_STATION
-// void HandleWFXSystemEvent(wfx_event_base_t eventBase, sl_wfx_generic_message_t * eventData);
-#endif
-
 namespace chip {
 namespace DeviceLayer {
